Struct normpath::BasePath

#[repr(transparent)]
pub struct BasePath(/* private fields */);

A borrowed path that has a prefix on Windows.

Note that comparison traits such as PartialEq will compare paths literally instead of comparing components. The former is more efficient and easier to use correctly.

Safety

This type should not be used for memory safety, but implementations can panic if this path is missing a prefix on Windows. A safe new_unchecked method might be added later that can safely create invalid base paths.

Although this type is annotated with #[repr(transparent)], the inner representation is not stable. Transmuting between this type and any other causes immediate undefined behavior.

Implementations

impl BasePath

pub fn new<'a, P>(path: P) -> Result<Cow<'a, Self>>where P: Into<Cow<'a, Path>>,

Creates a new base path.

On Windows, if path is missing a prefix, it will be joined to the current directory.

Errors

Returns an error if reading the current directory fails.

Examples

use std::path::Path;

use normpath::BasePath;

if cfg!(windows) {
    let path = Path::new(r"X:\foo\bar");
    assert_eq!(path, *BasePath::new(path)?);

    assert!(BasePath::new(Path::new(r"foo\bar")).is_ok());
}

pub fn try_new<P>(path: &P) -> Result<&Self, MissingPrefixError>where P: AsRef<Path> + ?Sized,

Creates a new base path.

Errors

On Windows, returns an error if path is missing a prefix.

Examples

use std::path::Path;

use normpath::BasePath;

if cfg!(windows) {
    let path = r"X:\foo\bar";
    assert_eq!(Path::new(path), BasePath::try_new(path)?);

    assert!(BasePath::try_new(r"foo\bar").is_err());
}

pub fn as_os_str(&self) -> &OsStr

Returns a reference to the wrapped path as a platform string.

pub fn as_path(&self) -> &Path

Returns a reference to the wrapped path.

pub fn canonicalize(&self) -> Result<BasePathBuf>

Equivalent to Path::canonicalize.

pub fn components(&self) -> Components<'_>

Equivalent to Path::components.

pub fn ends_with<P>(&self, child: P) -> boolwhere P: AsRef<Path>,

Equivalent to Path::ends_with.

pub fn exists(&self) -> bool

Equivalent to Path::exists.

pub fn extension(&self) -> Option<&OsStr>

Equivalent to Path::extension.

pub fn file_name(&self) -> Option<&OsStr>

Equivalent to Path::file_name.

pub fn file_stem(&self) -> Option<&OsStr>

Equivalent to Path::file_stem.

pub fn has_root(&self) -> bool

Equivalent to Path::has_root.

pub fn is_absolute(&self) -> bool

Equivalent to Path::is_absolute.

pub fn is_dir(&self) -> bool

Equivalent to Path::is_dir.

pub fn is_file(&self) -> bool

Equivalent to Path::is_file.

pub fn is_relative(&self) -> bool

Equivalent to Path::is_relative.

pub fn join<P>(&self, path: P) -> BasePathBufwhere P: AsRef<Path>,

An improved version of Path::join that handles more edge cases.

For example, on Windows, leading . and .. components of path will be normalized if possible. If self is a verbatim path, it would be invalid to normalize them later.

You should still call normalize before parent to normalize some additional components.

Examples

use std::path::Path;

use normpath::BasePath;

if cfg!(windows) {
    assert_eq!(
        Path::new(r"\\?\foo\baz\test.rs"),
        BasePath::try_new(r"\\?\foo\bar")
            .unwrap()
            .join("../baz/test.rs"),
    );
}

pub fn metadata(&self) -> Result<Metadata>

Equivalent to Path::metadata.

pub fn normalize(&self) -> Result<BasePathBuf>

Equivalent to PathExt::normalize.

pub fn normalize_virtually(&self) -> Result<BasePathBuf>

Equivalent to PathExt::normalize_virtually.

pub fn parent(&self) -> Result<Option<&Self>, ParentError>

Returns this path without its last component.

Returns Ok(None) if the last component is Component::RootDir.

You should usually only call this method on normalized paths. They will prevent an unexpected path from being returned due to symlinks, and some . and .. components will be normalized.

Errors

Returns an error if the last component is not Component::Normal or Component::RootDir. To ignore this error, use parent_unchecked.

Examples

use std::path::Path;

use normpath::BasePath;

if cfg!(windows) {
    assert_eq!(
        Path::new(r"X:\foo"),
        BasePath::try_new(r"X:\foo\bar").unwrap().parent()?.unwrap(),
    );
}

pub fn parent_unchecked(&self) -> Option<&Self>

Equivalent to Path::parent.

It is usually better to use parent.

Examples

use std::path::Path;

use normpath::BasePath;

if cfg!(windows) {
    assert_eq!(
        Path::new(r"X:\foo"),
        BasePath::try_new(r"X:\foo\..")
            .unwrap()
            .parent_unchecked()
            .unwrap(),
    );
}

pub fn read_dir(&self) -> Result<ReadDir>

Equivalent to Path::read_dir.

pub fn read_link(&self) -> Result<PathBuf>

Equivalent to Path::read_link.

pub fn starts_with<P>(&self, base: P) -> boolwhere P: AsRef<Path>,

Equivalent to Path::starts_with.

pub fn symlink_metadata(&self) -> Result<Metadata>

Equivalent to Path::symlink_metadata.